add new rules

+++ validation/rules/Max.py
@@ -7,7 +7,7 @@ class Max(BaseRule):
     def validate(self, attribute: str, value: Any) -> bool:
         if isinstance(value, Union[int, float]):
             return value <= float(self.payload)
-        return len(value) <= int(self.payload)
+        return len(value) <= float(self.payload)
 
     def message(self, attribute: str) -> str:
         return f"{attribute} length is bigger than {self.payload}"
